move integration test to mgo sub package
