+++ test/integration/mgo/items_store.go
@@ -1,4 +1,4 @@
-package integration
+package mgo
 
 import (
 	"time"
