+++ test/integration/mgo/items_store_test.go
@@ -1,4 +1,4 @@
-package integration
+package mgo
 
 import (
 	"os"
